= SipParser =  * m_iUseCount 를 int8_t 타입으로 수정함

+++ SipParser/SipMessage.h
@@ -136,7 +136,7 @@ public:
 	bool						m_bUseCompact;
 
 	/** 객체 사용 개수 */
-	int							m_iUseCount;
+	int8_t					m_iUseCount;
 
 	int Parse( const char * pszText, int iTextLen );
 	int ToString( char * pszText, int iTextSize );
